+++ tg/src/lib.rs
@@ -3,5 +3,6 @@
 pub mod admin;
 pub mod auth;
 pub mod client;
+pub mod forum;
 pub mod model;
 pub mod session;